+++ src/MemoizedCombinedRegexp.php
@@ -5,38 +5,28 @@ declare(strict_types=1);
 namespace Yiisoft\Strings;
 
 /**
- * `CombinedRegexp` optimizes matching of multiple regular expressions.
- * Read more about the concept in
- * {@see https://nikic.github.io/2014/02/18/Fast-request-routing-using-regular-expressions.html}.
+ * MemoizedCombinedRegexp is a decorator for {@see AbstractCombinedRegexp} that caches results of
+ * - {@see AbstractCombinedRegexp::matches()}
+ * - {@see AbstractCombinedRegexp::getMatchingPattern()}
+ * - {@see AbstractCombinedRegexp::getMatchingPatternPosition()}.
  */
-final class MemoizedCombinedRegexp
+final class MemoizedCombinedRegexp extends AbstractCombinedRegexp
 {
-    private CombinedRegexp $combinedRegexp;
-
+    /**
+     * @var array<string, array{matches:bool, position?:int}>
+     */
     private array $results = [];
 
-    /**
-     * @param string[] $patterns Regular expressions to combine.
-     * @param string $flags Flags to apply to all regular expressions.
-     */
     public function __construct(
-        private array $patterns,
-        string $flags = ''
+        private AbstractCombinedRegexp $decorated,
     ) {
-        $this->combinedRegexp = new CombinedRegexp($patterns, $flags);
     }
 
-    /**
-     * @return string The compiled pattern.
-     */
     public function getCompiledPattern(): string
     {
-        return $this->combinedRegexp->getCompiledPattern();
+        return $this->decorated->getCompiledPattern();
     }
 
-    /**
-     * Returns `true` whether the given string matches any of the patterns, `false` - otherwise.
-     */
     public function matches(string $string): bool
     {
         $this->evaluate($string);
@@ -44,46 +34,42 @@ final class MemoizedCombinedRegexp
         return $this->results[$string]['matches'];
     }
 
-    /**
-     * Returns pattern that matches the given string.
-     * @throws \Exception if the string does not match any of the patterns.
-     */
     public function getMatchingPattern(string $string): string
     {
         $this->evaluate($string);
 
-        return $this->patterns[$this->getMatchingPatternPosition($string)];
+        return $this->getPatterns()[$this->getMatchingPatternPosition($string)];
     }
 
-    /**
-     * Returns position of the pattern that matches the given string.
-     * @throws \Exception if the string does not match any of the patterns.
-     */
     public function getMatchingPatternPosition(string $string): int
     {
         $this->evaluate($string);
 
-        return $this->results[$string]['position'] ?? throw new \Exception(
-            sprintf(
-                'Failed to match pattern "%s" with string "%s".',
-                $this->getCompiledPattern(),
-                $string,
-            )
-        );
+        return $this->results[$string]['position'] ?? $this->throwFailedMatchException($string);
     }
 
-    protected function evaluate(string $string): void
+    private function evaluate(string $string): void
     {
         if (isset($this->results[$string])) {
             return;
         }
         try {
-            $position = $this->combinedRegexp->getMatchingPatternPosition($string);
+            $position = $this->decorated->getMatchingPatternPosition($string);
 
             $this->results[$string]['matches'] = true;
             $this->results[$string]['position'] = $position;
         } catch (\Exception) {
             $this->results[$string]['matches'] = false;
         }
+    }
+
+    public function getPatterns(): array
+    {
+        return $this->decorated->getPatterns();
+    }
+
+    public function getFlags(): string
+    {
+        return $this->decorated->getFlags();
     }
 }