+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2.2.0 under development
 
+- New #103: Add `MemoizedCombinedRegexp` class (@xepozz)
 - New #102, #106: Add `CombinedRegexp` class (@xepozz, @vjik)
 - Enh #106: Using fully-qualified function calls to improve performance (@vjik)